Refactor/Simplify steps logic
