+++ Trente/Localizable.xcstrings
@@ -76,6 +76,9 @@
     "Clothes" : {
 
     },
+    "Create" : {
+
+    },
     "CURRENT MONTH" : {
 
     },
@@ -129,9 +132,6 @@
 
     },
     "Graph Card" : {
-
-    },
-    "Hello here" : {
 
     },
     "How would you like to split this income?" : {
@@ -204,9 +204,6 @@
 
     },
     "Previous" : {
-
-    },
-    "Recurrence" : {
 
     },
     "Recurring transaction" : {
